--- src/main/java/duke/command/CommandWord.java
@@ -1,8 +1,0 @@
-package duke.command;
-
-/**
- * Command words that can be ysed
- */
-public enum CommandWord {
-    TODO, DEADLINE, EVENT, MARK, UNMARK, DELETE, FIND
-}
