--- src/main/java/duke/command/ExitCommand.java
@@ -1,19 +1,0 @@
-package duke.command;
-
-import duke.storage.Storage;
-import duke.task.TaskList;
-import duke.ui.Ui;
-
-/**
- * Represents a command to exit the Duke program.
- */
-public class ExitCommand extends Command {
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.exit();
-    }
-    @Override
-    public boolean isExit() {
-        return true;
-    }
-}
