--- src/main/java/duke/command/ListCommand.java
@@ -1,20 +1,0 @@
-package duke.command;
-
-import duke.storage.Storage;
-import duke.task.TaskList;
-import duke.ui.Ui;
-
-/**
- * Represents a command to list all tasks in the task list.
- */
-public class ListCommand extends Command {
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.showListMessage(tasks.getSize());
-        tasks.list();
-    }
-    @Override
-    public boolean isExit() {
-        return false;
-    }
-}
