--- src/main/java/duke/task/TaskType.java
@@ -1,8 +1,0 @@
-package duke.task;
-
-/**
- * The types of tasks that can be used.
- */
-public enum TaskType {
-    TODO, DEADLINE, EVENT
-}
