+++ src/main/java/mort/Launcher.java
@@ -1,0 +1,10 @@
+package mort;
+
+import javafx.application.Application;
+import mort.ui.Main;
+
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(Main.class, args);
+    }
+}
