+++ src/main/java/mort/command/CommandWord.java
@@ -1,0 +1,8 @@
+package mort.command;
+
+/**
+ * Command words that can be ysed
+ */
+public enum CommandWord {
+    TODO, DEADLINE, EVENT, MARK, UNMARK, DELETE, FIND
+}
