+++ src/main/java/mort/command/FindCommand.java
@@ -1,0 +1,24 @@
+package mort.command;
+
+import mort.storage.Storage;
+import mort.task.TaskList;
+import mort.ui.Ui;
+
+/**
+ * Represents a command to find a task in the task list.
+ */
+public class FindCommand extends Command {
+    /** Keyword to be searched */
+    private String keyword;
+    public FindCommand(String keyword) {
+        this.keyword = keyword;
+    }
+    @Override
+    public String execute(TaskList tasks, Ui ui, Storage storage) {
+        return ui.getFindMessage(this.keyword, tasks.find(this.keyword));
+    }
+    @Override
+    public boolean isExit() {
+        return false;
+    }
+}
