+++ src/main/java/mort/command/ListCommand.java
@@ -1,0 +1,19 @@
+package mort.command;
+
+import mort.storage.Storage;
+import mort.task.TaskList;
+import mort.ui.Ui;
+
+/**
+ * Represents a command to list all tasks in the task list.
+ */
+public class ListCommand extends Command {
+    @Override
+    public String execute(TaskList tasks, Ui ui, Storage storage) {
+        return ui.getListMessage(tasks.getSize()) + tasks.list();
+    }
+    @Override
+    public boolean isExit() {
+        return false;
+    }
+}
