+++ src/main/java/mort/task/TaskType.java
@@ -1,0 +1,8 @@
+package mort.task;
+
+/**
+ * The types of tasks that can be used.
+ */
+public enum TaskType {
+    TODO, DEADLINE, EVENT
+}
