+++ src/main/resources/view/DialogBox.fxml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.text.Font?>
+
+<fx:root alignment="TOP_RIGHT" minHeight="-Infinity" maxHeight="1.7976931348623157E308"
+         maxWidth="1.7976931348623157E308" prefWidth="500.0" type="javafx.scene.layout.HBox"
+         xmlns="http://javafx.com/javafx/18" xmlns:fx="http://javafx.com/fxml/1">
+    <children>
+        <Label fx:id="dialog" text="Label" wrapText="true">
+            <padding>
+                <Insets left="10.0" top="5.0" bottom="5.0"/>
+            </padding>
+            <font>
+                <Font name="Baskerville" size="15.0" />
+            </font>
+        </Label>
+        <ImageView fx:id="displayPicture" fitHeight="99.0" fitWidth="99.0" pickOnBounds="true" preserveRatio="true" />
+    </children>
+    <padding>
+        <Insets bottom="10.0" left="2.5" right="5.0" top="10.0" />
+    </padding>
+</fx:root>
