+++ src/main/resources/view/MainWindow.fxml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.ScrollPane?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.text.Font?>
+
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity"
+            prefHeight="800.0" prefWidth="500.0" xmlns="http://javafx.com/javafx/18"
+            xmlns:fx="http://javafx.com/fxml/1" fx:controller="mort.ui.MainWindow">
+    <children>
+        <TextField fx:id="userInput" layoutY="758.0" onAction="#handleUserInput" prefHeight="41.0"
+                   prefWidth="424.0" promptText="Enter a command" AnchorPane.bottomAnchor="1.0">
+            <font>
+                <Font name="Helvetica Neue" size="13.0" />
+            </font>
+        </TextField>
+        <Button fx:id="sendButton" layoutX="424.0" layoutY="758.0" mnemonicParsing="false"
+                onAction="#handleUserInput" prefHeight="41.0" prefWidth="76.0" text="Send">
+            <font>
+                <Font name="Helvetica Neue" size="14.0" />
+             </font>
+        </Button>
+        <ScrollPane fx:id="scrollPane" hbarPolicy="NEVER" hvalue="1.0" prefHeight="757.0" prefWidth="500.0" vvalue="1.0">
+            <content>
+                <VBox fx:id="dialogContainer" prefHeight="752.0" prefWidth="488.0" />
+            </content>
+        </ScrollPane>
+    </children>
+</AnchorPane>
